procrastinating on christmas eve by doing little changes to remove some warning messages during compile

+++ fittino.cpp
@@ -494,8 +494,8 @@ Fittino::Fittino(const Input* input)
       histo->SetBinContent( 9,  meff9 * norm);
       histo->SetBinContent(10, meff10 * norm);
       
-      p.first = m0;
-      p.second = m12;
+      p.first = (int)m0;
+      p.second = (int)m12;
       signalXsec[p] = histo;
     }
